Create README.md
I was the administrator of a Electronic Managing documents software : Isotarget developed by Gedys. I didn't not made any development, I only made the specifical configuration for the company where I work accoding to our needs. It was easy and I didn't need to be ingeneer or developer to do that. Know I would like to progress in this topic.